+++ relays/bin-substrate/src/cli/relay_headers.rs
@@ -59,12 +59,14 @@ macro_rules! select_bridge {
 				type Source = relay_rialto_client::Rialto;
 				type Target = relay_millau_client::Millau;
 				type Finality = crate::rialto_millau::rialto_headers_to_millau::RialtoFinalityToMillau;
+
 				$generic
 			}
 			RelayHeadersBridge::WestendToMillau => {
 				type Source = relay_westend_client::Westend;
 				type Target = relay_millau_client::Millau;
 				type Finality = crate::rialto_millau::westend_headers_to_millau::WestendFinalityToMillau;
+
 				$generic
 			}
 		}
